Allow one OLD process to serve multiple databases
Makes modifications to the database session logic and the routing logic so that
one OLD process can serve different logical OLD instances, depending on the
first element of the URL path it receives. For example, requests to
/blaold/forms/ might return all forms for the Blackfoot OLD while those to
/okaold/foms/ might return all those for the Okanagan OLD. This change will
allow the OLD to be stateless and paves the way for a containerized,
load-balanced, 12-factor type of deploy where multiple OLD processes can share
the load of responding to requests. This change required non-trivial
modifications to how SQLAlchemy database sessions are created and shared. In
short, we are no longer using the transaction context manager but are manually
closing db sessions using an end-of-request event.

Several warnings were also dealt with: cases where file descriptors were
created (i.e., calls to `open`) without context managers were fixed; calls to
the deprecated formencode methods `_to_python` and `validate_python` were
replaced with their recommended counterparts; and user salts are now being
passed to SQLAlchemy as Unicode strings.

Also fixed a bug in the logic for when a morphoparser recognizes itself as
having changed such that its cache of transcriptions-to-parses has been
invalidated.